+++ arviz/data/base.py
@@ -157,7 +157,11 @@ def generate_dims_coords(
         dim_name = dims[idx]
         if dim_name not in coords:
             coords[dim_name] = np.arange(index_origin, dim_len + index_origin)
-    coords = {key: coord for key, coord in coords.items() if any(key == dim for dim in dims)}
+    coords = {
+        key: coord
+        for key, coord in coords.items()
+        if any(key == dim for dim in dims + default_dims)
+    }
     return dims, coords
 
 
@@ -246,7 +250,7 @@ def numpy_to_data_array(
         coords["draw"] = np.arange(index_origin, n_samples + index_origin)
 
     # filter coords based on the dims
-    coords = {key: xr.IndexVariable((key,), data=coords[key]) for key in dims}
+    coords = {key: xr.IndexVariable((key,), data=np.asarray(coords[key])) for key in dims}
     return xr.DataArray(ary, coords=coords, dims=dims)