+++ arviz/tests/base_tests/test_data.py
@@ -26,6 +26,7 @@ from ... import (
     list_datasets,
     load_arviz_data,
     to_netcdf,
+    extract_dataset,
 )
 
 from ...data.base import dict_to_dataset, generate_dims_coords, infer_stan_dtypes, make_attrs
@@ -147,6 +148,26 @@ def test_dims_coords():
     assert len(coords["x_dim_0"]) == 4
     assert len(coords["x_dim_1"]) == 20
     assert len(coords["x_dim_2"]) == 5
+
+
+def test_dims_coords_default_dims():
+    shape = 4, 7
+    var_name = "x"
+    dims, coords = generate_dims_coords(
+        shape,
+        var_name,
+        dims=["dim1", "dim2"],
+        coords={"chain": ["a", "b", "c"]},
+        default_dims=["chain", "draw"],
+    )
+    assert "dim1" in dims
+    assert "dim2" in dims
+    assert "chain" not in dims
+    assert "draw" not in dims
+    assert len(coords["dim1"]) == 4
+    assert len(coords["dim2"]) == 7
+    assert len(coords["chain"]) == 3
+    assert "draw" not in coords
 
 
 def test_dims_coords_extra_dims():
@@ -946,6 +967,16 @@ def test_dict_to_dataset_event_dims_error():
         convert_to_dataset(datadict, coords=coords, dims={"a": ["b", "c"]})
 
 
+def test_dict_to_dataset_with_tuple_coord():
+    datadict = {"a": np.random.randn(100), "b": np.random.randn(1, 100, 10)}
+    dataset = convert_to_dataset(datadict, coords={"c": tuple(range(10))}, dims={"b": ["c"]})
+    assert set(dataset.data_vars) == {"a", "b"}
+    assert set(dataset.coords) == {"chain", "draw", "c"}
+
+    assert set(dataset.a.coords) == {"chain", "draw"}
+    assert set(dataset.b.coords) == {"chain", "draw", "c"}
+
+
 def test_convert_to_dataset_idempotent():
     first = convert_to_dataset(np.random.randn(100))
     second = convert_to_dataset(first)
@@ -1387,3 +1418,38 @@ class TestDataArrayToDataset:
         assert inference_data.prior.chain.shape == shape[:1]
         assert inference_data.prior.draw.shape == shape[1:2]
         assert inference_data.prior[var_name].shape == shape
+
+
+class TestExtractDataset:
+    def test_default(self):
+        idata = load_arviz_data("centered_eight")
+        post = extract_dataset(idata)
+        assert isinstance(post, xr.Dataset)
+        assert "sample" in post.dims
+        assert post.theta.size == (4 * 500 * 8)
+
+    def test_seed(self):
+        idata = load_arviz_data("centered_eight")
+        post = extract_dataset(idata, rng=7)
+        post_pred = extract_dataset(idata, group="posterior_predictive", rng=7)
+        assert all(post.sample == post_pred.sample)
+
+    def test_no_combine(self):
+        idata = load_arviz_data("centered_eight")
+        post = extract_dataset(idata, combined=False)
+        assert "sample" not in post.dims
+        assert post.dims["chain"] == 4
+        assert post.dims["draw"] == 500
+
+    def test_var_name_group(self):
+        idata = load_arviz_data("centered_eight")
+        prior = extract_dataset(idata, group="prior", var_names="the", filter_vars="like")
+        assert prior.attrs == idata.prior.attrs
+        assert "theta" in prior.data_vars
+        assert "mu" not in prior.data_vars
+
+    def test_subset_samples(self):
+        idata = load_arviz_data("centered_eight")
+        post = extract_dataset(idata, num_samples=10)
+        assert post.dims["sample"] == 10
+        assert post.attrs == idata.posterior.attrs